replace login modal navigation with render
